refactor: Move automation e2e tests to `/test` directory
