+++ test/automation/automation_integration_test.go
@@ -2,7 +2,7 @@
 
 /*
  * @license
- * Copyright 2023 Dynatrace LLC
+ * Copyright 2025 Dynatrace LLC
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package v2
+package automation
 
 import (
 	"fmt"
@@ -28,26 +28,25 @@ import (
 
 	"github.com/dynatrace/dynatrace-configuration-as-code/v2/cmd/monaco/integrationtest"
 	"github.com/dynatrace/dynatrace-configuration-as-code/v2/cmd/monaco/integrationtest/utils/monaco"
+	"github.com/dynatrace/dynatrace-configuration-as-code/v2/cmd/monaco/integrationtest/v2"
 )
 
 // tests all configs for a single environment
 func TestIntegrationAutomation(t *testing.T) {
 
-	configFolder := "test-resources/integration-automation/"
+	configFolder := "testdata/"
 	manifest := configFolder + "manifest.yaml"
 
 	envs := map[string]string{}
-	if isHardeningEnvironment() {
+	if v2.IsHardeningEnvironment() {
 		envs["WORKFLOW_ACTOR"] = os.Getenv("WORKFLOW_ACTOR")
 	}
 
-	Run(t, configFolder,
-		Options{
-			WithManifestPath(manifest),
-			WithSuffix("Automation"),
-			WithEnvVars(envs),
+	v2.Run(t, configFolder,
+		v2.Options{
+			v2.WithEnvVars(envs),
 		},
-		func(fs afero.Fs, _ TestContext) {
+		func(fs afero.Fs, _ v2.TestContext) {
 			// This causes Creation of all automation objects
 			err := monaco.Run(t, fs, fmt.Sprintf("monaco deploy %s --verbose", manifest))
 			assert.NoError(t, err)
