+++ test/automation/testdata/manifest.yaml
