+++ test/automation/testdata/project/combined/business-calendar.json
