+++ test/automation/testdata/project/combined/config.yaml
