+++ test/automation/testdata/project/combined/jira-issue-workflow.json
