+++ test/automation/testdata/project/combined/scheduling-rule.json
