+++ test/automation/testdata/project/common/script.js
