+++ test/automation/testdata/project/scheduling-rule/02c70740-74f9-422f-b325-073f0f656d84.json
