+++ test/automation/testdata/project/scheduling-rule/02cac1f9-7a7e-4939-b8e1-77e8f3a820ef.json
