+++ test/automation/testdata/project/scheduling-rule/0c18677b-f24c-4fa2-95a3-1ffee75d763b.json
