+++ test/automation/testdata/project/scheduling-rule/config.yaml
